+++ app/src/main/java/com/leeweeder/weighttracker/ui/util/Util.kt
@@ -8,7 +8,7 @@ import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 import java.util.Locale
 
-fun Double.formatToOneDecimalPlace(showTrailingZero: Boolean = true, showPlusSign: Boolean = false): String {
+fun Float.formatToOneDecimalPlace(showTrailingZero: Boolean = true, showPlusSign: Boolean = false): String {
     val formatted = String.format(locale = Locale.getDefault(), format = "%.1f", this)
     if (!showTrailingZero) {
         return formatted.trimEnd('0').trimEnd('.')
